chore: remove unused code and rename dir
- Remove code replaced by new DataStore, DataType and IndexWriter
  implementations
- Skip failing tests for now
- Move all code to `src` folder to prepare for compiling typescript
  declaration files for package publishing
