+++ src/blob-store/live-download.js
@@ -36,10 +36,10 @@ export class LiveDownload extends TypedEmitter {
    * @param {Iterable<import('hyperdrive')>} drives
    * @param {import('./index.js').InternalDriveEmitter} emitter
    * @param {object} options
-   * @param {import('../types').BlobFilter} [options.filter] Filter blobs of specific types and/or sizes to download
+   * @param {import('../types.js').BlobFilter} [options.filter] Filter blobs of specific types and/or sizes to download
    * @param {AbortSignal} [options.signal]
    */
-  constructor (drives, emitter, { filter, signal }) {
+  constructor(drives, emitter, { filter, signal }) {
     super()
     this.#signal = signal
 
@@ -48,10 +48,10 @@ export class LiveDownload extends TypedEmitter {
     }
 
     /** @param {import('hyperdrive')} drive */
-    const addDrive = drive => {
+    const addDrive = (drive) => {
       const download = new DriveLiveDownload(drive, {
         filter,
-        signal
+        signal,
       })
       this.#driveLiveDownloads.add(download)
       download.on('state', emitState)
@@ -75,7 +75,7 @@ export class LiveDownload extends TypedEmitter {
   /**
    * @returns {BlobDownloadState | BlobDownloadStateError}
    */
-  get state () {
+  get state() {
     return combineStates(this.#driveLiveDownloads, { signal: this.#signal })
   }
 }
@@ -101,10 +101,10 @@ export class DriveLiveDownload extends TypedEmitter {
    * Like drive.download() but 'live',
    * @param {import('hyperdrive')} drive
    * @param {object} options
-   * @param {import('../types').BlobFilter} [options.filter] Filter blobs of specific types and/or sizes to download
+   * @param {import('../types.js').BlobFilter} [options.filter] Filter blobs of specific types and/or sizes to download
    * @param {AbortSignal} [options.signal]
    */
-  constructor (drive, { filter, signal } = {}) {
+  constructor(drive, { filter, signal } = {}) {
     super()
     this.#drive = drive
     this.#folders = filterToFolders(filter)
@@ -126,7 +126,7 @@ export class DriveLiveDownload extends TypedEmitter {
   /**
    * @returns {BlobDownloadState | BlobDownloadStateError}
    */
-  get state () {
+  get state() {
     if (this.#error)
       return {
         haveCount: this.#haveCount,
@@ -134,7 +134,7 @@ export class DriveLiveDownload extends TypedEmitter {
         wantCount: this.#downloads.size,
         wantBytes: this.#wantBytes,
         error: this.#error,
-        status: 'error'
+        status: 'error',
       }
     return {
       haveCount: this.#haveCount,
@@ -148,11 +148,11 @@ export class DriveLiveDownload extends TypedEmitter {
         ? 'checking'
         : this.#downloads.size > 0
         ? 'downloading'
-        : 'downloaded'
+        : 'downloaded',
     }
   }
 
-  async #start () {
+  async #start() {
     const blobsCore = await this.#getBlobsCore()
     /* c8 ignore next */
     if (this.#signal?.aborted || !blobsCore) return // Can't get here in tests
@@ -164,7 +164,7 @@ export class DriveLiveDownload extends TypedEmitter {
       const entryStream = this.#drive.list(folder, { recursive: true })
       if (this.#signal) {
         this.#signal.addEventListener('abort', () => entryStream.destroy(), {
-          once: true
+          once: true,
         })
       }
       for await (const entry of entryStream) {
@@ -186,10 +186,14 @@ export class DriveLiveDownload extends TypedEmitter {
     // edits of blobs, so this should not be an issue. `keyEncoding` is
     // necessary because hyperdrive stores file index data under the `files`
     // sub-encoding key
-    const historyStream = bee.createHistoryStream({ live: true, gt: seq, keyEncoding })
+    const historyStream = bee.createHistoryStream({
+      live: true,
+      gt: seq,
+      keyEncoding,
+    })
     if (this.#signal) {
       this.#signal.addEventListener('abort', () => historyStream.destroy(), {
-        once: true
+        once: true,
       })
     }
     for await (const entry of historyStream) {
@@ -221,7 +225,7 @@ export class DriveLiveDownload extends TypedEmitter {
    *
    * @returns {Promise<import('hypercore').default | undefined>}
    */
-  async #getBlobsCore () {
+  async #getBlobsCore() {
     if (this.#drive.blobs) return this.#drive.blobs.core
     await this.#drive.ready()
     await this.#drive.core.update({ wait: true })
@@ -243,7 +247,7 @@ export class DriveLiveDownload extends TypedEmitter {
   }
 
   /** @param {Error} e */
-  #handleError (e) {
+  #handleError(e) {
     this.#error = e
     this.emit('state', this.state)
   }
@@ -254,7 +258,7 @@ export class DriveLiveDownload extends TypedEmitter {
    * @param {import('hypercore').default} core
    * @param {{ blockOffset: number, blockLength: number, byteLength: number }} blob
    */
-  async #processEntry (
+  async #processEntry(
     core,
     { blockOffset: start, blockLength: length, byteLength }
   ) {
@@ -289,7 +293,7 @@ export class DriveLiveDownload extends TypedEmitter {
  * @param {{ signal?: AbortSignal }} options
  * @returns
  */
-export function combineStates (liveDownloads, { signal } = {}) {
+export function combineStates(liveDownloads, { signal } = {}) {
   /** @type {BlobDownloadState | BlobDownloadStateError} */
   let combinedState = {
     haveCount: 0,
@@ -297,7 +301,7 @@ export function combineStates (liveDownloads, { signal } = {}) {
     wantCount: 0,
     wantBytes: 0,
     error: null,
-    status: 'downloaded'
+    status: 'downloaded',
   }
   for (const { state } of liveDownloads) {
     combinedState.haveCount += state.haveCount
@@ -329,15 +333,15 @@ export function combineStates (liveDownloads, { signal } = {}) {
 /**
  * Convert a filter to an array of folders that need to be downloaded
  *
- * @param {import('../types').BlobFilter} [filter]
+ * @param {import('../types.js').BlobFilter} [filter]
  * @returns {string[]} array of folders that match the filter
  */
-function filterToFolders (filter) {
+function filterToFolders(filter) {
   if (!filter) return ['/']
   const folders = []
   for (const [
     type,
-    variants
+    variants,
   ] of /** @type {import('type-fest').Entries<typeof filter>} */ (
     Object.entries(filter)
   )) {
@@ -356,14 +360,14 @@ function filterToFolders (filter) {
  * @param {string[]} folders
  * @returns {boolean}
  */
-function matchesFolder (path, folders) {
+function matchesFolder(path, folders) {
   for (const folder of folders) {
     if (path.startsWith(folder)) return true
   }
   return false
 }
 
-/** @param {Pick<import('../types').BlobId, 'type' | 'variant'>} opts */
-function makePath ({ type, variant }) {
+/** @param {Pick<import('../types.js').BlobId, 'type' | 'variant'>} opts */
+function makePath({ type, variant }) {
   return `/${type}/${variant}`
 }
