+++ src/core-manager/core-index.js
