+++ src/core-manager/index.js
