+++ src/core-manager/messages.js
