+++ src/core-manager/random-access-file-pool.js
