+++ src/core-manager/replication-state-machine.js
