+++ src/index.js
@@ -1,0 +1,1 @@
+export { MapeoProject } from './mapeo-project.js'
