+++ src/mapeo-project.js
@@ -3,9 +3,9 @@ import { drizzle } from 'drizzle-orm/better-sqlite3'
 import { migrate } from 'drizzle-orm/better-sqlite3/migrator'
 
 import { CoreManager } from './core-manager/index.js'
-import { DataStore } from './datastore/data-store-new.js'
-import { DataType } from './datatype/data-type-new.js'
-import { IndexWriter } from './index-writer.js'
+import { DataStore } from './datastore/index.js'
+import { DataType } from './datatype/index.js'
+import { IndexWriter } from './index-writer/index.js'
 import { observationTable } from './schema/project.js'
 import RandomAccessFile from 'random-access-file'
 import RAM from 'random-access-memory'
